First go at replacing some of the more overly ✨abstracted✨ azure functions
