--- packages/functions/add-organisation-to-user/function.json
@@ -1,21 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "function",
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req",
-      "methods": [
-        "get",
-        "post"
-      ]
-    },
-    {
-      "type": "http",
-      "direction": "out",
-      "name": "res"
-    }
-  ],
-  "entryPoint": "AddOrganisationToUser",
-  "scriptFile": "../dist/add-organisation-to-user/index.js"
-}
